Extract modules to src package
